% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_benchmark.R
\name{add_benchmark}
\alias{add_benchmark}
\alias{add_benchmark.default}
\alias{add_benchmark.sensemakr}
\title{Add benchmarks to sensemakr object}
\usage{
add_benchmark(obj, variables)

\method{add_benchmark}{default}(obj, variables)

\method{add_benchmark}{sensemakr}(obj, variables)
}
\arguments{
\item{obj}{A `sensemakr` object to add benchmarks to.}

\item{variables}{A vector of character strings describing the variable or
variables to add as benchmarks}
}
\value{
The `sensemakr` object `x` with benchmarks added.
}
\description{
This function adds a benchmark variable to a `sensemakr` object. Benchmark
variables are variables which exist in the real models fit, but which stand
in for hypothetical unobserved variables and can attenuate observed effects
for variables of interest. Our suggestion is that subject-matter knowledge
guides the process of selecting a benchmark variable based on a belief that
it is the largest observed confounder.
}
