% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_plot.R
\name{contour_plot}
\alias{contour_plot}
\title{Contour plot}
\usage{
contour_plot(...)
}
\arguments{
\item{...}{Arguments which will pass through to the methods invoked.}
}
\description{
`contour_plot` is a generic function to produce sensitivity analysis
contour plots. The function invokes particular methods which depend on
the class of the first argument. For documentation on using  `contour_plot`
with `sensemakr` objects, check \link{contour_plot.sensemakr}. For
documentation on using `contour_plot` with direct numeric arguments, check
\link{contour_plot.default}
}
