% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_plot.sensemakr.R
\name{contour_plot.sensemakr}
\alias{contour_plot.sensemakr}
\title{Contour plot.}
\usage{
\method{contour_plot}{sensemakr}(x, ..., lim = c(0, 0.4, 0.001),
  nlevels = 20, plot_t = FALSE, col.contour = "grey40",
  col.line = "red", multipliers_y = c(1, 2, 3),
  multipliers_d = multipliers_y)
}
\arguments{
\item{x}{A `sensemakr`  object to produce a contour plot from.}

\item{...}{Additional plotting graphical parameters}

\item{lim}{A numeric vector of length 3 containing the beginning, end, and
increment of a sequence, forms the axes of the bound plot}

\item{nlevels}{How many contour lines to produce}

\item{plot_t}{Logical, default FALSE describing whether to plot the
t-statistic (TRUE) or effect estimate (FALSE).}

\item{col.contour}{A color parameter for contour plot lines}

\item{col.line}{A color parameter for the threshold / critical value line}

\item{multipliers_y}{A vector of multipliers, k, describing how many times
as strong the hypothetical confounders are than the benchmark confounder
with respect to the outcome. Only usable with `sensemakr` objects that
contain benchmark variables.}

\item{multipliers_d}{A vector of multipliers, k, describing how many times
as strong the hypothetical confounders are than the benchmark confounder
with respect to the treatment. Only usable with `sensemakr` objects that
contain benchmark variables.}
}
\description{
Produces a sensitivity analysis contour plot. The produced plot maps how
observed effects could attenuate in the presence of an unobserved confounder
that is correlated with treatment and outcome.
}
