% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{darfur}
\alias{darfur}
\title{Post-conflict Darfur data}
\format{A data frame with 1276 rows and 14 columns.}
\usage{
darfur
}
\description{
Data on attitudes of Darfurian refugees in eastern Chad. The key "treatment"
variable is \code{directharm}, which indicates that individual was physically
injured during attacks on villages in Darfur, largely between 2003 and 2004.

Key  covariates include \code{herder_dar}
(whether they were a herder in Darfur), \code{farmer_dar} (whether they were a
farmer in Darfur), \code{age}, \code{female} (indicator for female), and
\code{past_voted} (whether they report having voted in an earlier election,
prior to the conflict).
}
\references{
Cinelli, C. and Hazlett, C. "Making Sense of Sensitivity: Extending Omitted Variable Bias." (2018)

Hazlett, C. (2013). Angry or weary? The effect of personal violence on attitudes towards peace in darfur. Working Paper.
}
\keyword{datasets}
