% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_stats.R
\name{error_if_no_dof.lm}
\alias{error_if_no_dof.lm}
\title{Helper function for checking dof statistics}
\usage{
\method{error_if_no_dof}{lm}(model, ...)
}
\arguments{
\item{model}{model to check dof in}

\item{...}{arguments passed to other methods.}
}
\description{
This is an internal function used for checking dof statistics.
}
\keyword{internal}
