% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_stats.R
\name{error_if_no_dof}
\alias{error_if_no_dof}
\title{Helper function for checking dof statistics}
\usage{
error_if_no_dof(...)
}
\arguments{
\item{...}{arguments passed to other methods.}
}
\description{
This is an internal function used for checking dof statistics.
}
