% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_plot.R
\name{extreme_plot}
\alias{extreme_plot}
\title{Extreme scenario plots}
\usage{
extreme_plot(...)
}
\arguments{
\item{...}{Arguments which will pass through to the methods invoked.}
}
\description{
`extreme_plot` is a generic function to produce extreme scenario plots. The
function invokes particular methods which depend on the class of the first
argument. For documentation on using  `extreme_plot` with `sensemakr`
objects, check \link{extreme_plot.sensemakr}. For documentation on using
`extreme_plot` with direct numeric arguments, check
\link{extreme_plot.default}
}
