% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensemakr.R
\name{get_RV}
\alias{get_RV}
\title{Obtains the robustness value at level q}
\usage{
get_RV(r2_yd, q = 1)
}
\arguments{
\item{q}{the q value: we are concerned with bias of 100q% of the original estimate}

\item{R2_yd}{Partial R2 of the treatment explaining the outcome.}
}
