% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensemakr.R
\name{getbiasR2}
\alias{getbiasR2}
\title{Computes  bias caused by unobserved confounder}
\usage{
getbiasR2(se, df, r2d, r2y, ...)
}
\arguments{
\item{se}{standard error of original  treatment effect estimate}

\item{df}{degrees of freedom of the original linear model}

\item{r2d}{hypothetical partial R2 of the confounder with the treatment}

\item{r2y}{hypothetical partial R2 of the confounder with the outcome}
}
\description{
This function computes the bias caused by an unobserved confounder
with certain characteristics.
}
