% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensemakr.R
\name{groupR2}
\alias{groupR2}
\title{The function '?groupR2()'}
\usage{
groupR2(model, terms_4_group)
}
\arguments{
\item{model}{an 'lm' object}

\item{terms_4_group}{a character vector of one or more terms to be viewed together as a single group.
NOTE: The elements in 'terms_4_group' must match the character values of "(attr(terms(formula(model)),'term.labels'))".}
}
\value{
a numeric scalar representing the R2 value of with-holding the model matrix columns associated with 'terms_4_group'
}
\description{
forms R2 quantities for a model.matrix column group tied to their model's term
}
\examples{
none, a low level helper
}
