% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{interpret}
\alias{interpret}
\title{Low level function to interpret sensemakr quantities}
\usage{
interpret(sensemakr, q = 1)
}
\arguments{
\item{sensemakr}{object with 'sensemakr' class.}

\item{q}{FOO}
}
\description{
provides text interpretation in the console
}
\examples{
# loads data
data("darfur")

# fits model
model  = lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
               pastvoted + hhsize_darfur + female + village, data = darfur)

# runs benchmarking etc
sense = sensemakr(model, treatment = "directlyharmed")

interpret(sense, q = 0.5)
interpret(sense, q = 0.6)
}
