% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\alias{make_table.default}
\alias{make_table.sensemakr}
\title{Produce a LaTeX summary of sensemakr output}
\usage{
make_table(x, ...)

\method{make_table}{default}(x, ...)

\method{make_table}{sensemakr}(x, digits = 3, ...)
}
\arguments{
\item{x}{A `sensemakr` object}

\item{...}{Additional display parameters governing the LaTeX table. See
description below for details.}

\item{digits}{A number indicating how many digits to display (percentage
outputs are rounded before multiplying by 100)}
}
\description{
This function produces a LaTeX summary of `sensemakr` object output: it
briefly summarizes the observed effect, standard error, partial R^2,
and robustness values as well as partial R^2 for variables used as
benchmarks. To output this table directly via RMarkdown / knitr, please be
sure to use the 'results = "asis"' chunk option.
}
\details{
Additional display parameters include `outcome_label`, which overrides the
outcome variable label, `treatment_label`, which overrides the treatment
variable label, and `benchmark_label`, which overrides the benchmark
variable label(s), `caption`, which adds a caption to the bottom of the
LaTeX table, and `label`, which adds a label to the bottom of the LaTeX
table.
}
