% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_stats.R
\name{model_helper}
\alias{model_helper}
\title{Helper function for extracting model statistics}
\usage{
model_helper(model, covariates = NULL, ...)
}
\arguments{
\item{model}{model to extract statistics from}

\item{covariates}{model covariates from which statistics will be extracted.}
}
\description{
This is an internal function used for extracting the necessary statistics from the models.
}
