% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovb_plots.R
\name{ovb_extreme_plot}
\alias{ovb_extreme_plot}
\alias{ovb_extreme_plot.lm}
\alias{ovb_extreme_plot.fixest}
\alias{ovb_extreme_plot.formula}
\alias{ovb_extreme_plot.numeric}
\title{Extreme scenarios plots of omitted variable bias}
\usage{
ovb_extreme_plot(...)

\method{ovb_extreme_plot}{lm}(
  model,
  treatment,
  benchmark_covariates = NULL,
  kd = 1,
  r2yz.dx = c(1, 0.75, 0.5),
  r2dz.x = NULL,
  reduce = TRUE,
  threshold = 0,
  lim = min(max(c(r2dz.x, 0.4), na.rm = TRUE) + 0.1, 0.5),
  legend = TRUE,
  cex.legend = 0.65,
  legend.bty = "n",
  ...
)

\method{ovb_extreme_plot}{fixest}(
  model,
  treatment,
  benchmark_covariates = NULL,
  kd = 1,
  r2yz.dx = c(1, 0.75, 0.5),
  r2dz.x = NULL,
  reduce = TRUE,
  threshold = 0,
  lim = min(max(c(r2dz.x, 0.4), na.rm = TRUE) + 0.1, 0.5),
  legend = TRUE,
  cex.legend = 0.65,
  legend.bty = "n",
  ...
)

\method{ovb_extreme_plot}{formula}(
  formula,
  method = c("lm", "feols"),
  vcov = "iid",
  data,
  treatment,
  benchmark_covariates = NULL,
  kd = 1,
  r2yz.dx = c(1, 0.75, 0.5),
  r2dz.x = NULL,
  reduce = TRUE,
  threshold = 0,
  lim = min(max(c(r2dz.x, 0.4), na.rm = TRUE) + 0.1, 0.5),
  legend = TRUE,
  cex.legend = 0.65,
  legend.bty = "n",
  ...
)

\method{ovb_extreme_plot}{numeric}(
  estimate,
  se,
  dof,
  r2dz.x = NULL,
  r2yz.dx = c(1, 0.75, 0.5),
  reduce = TRUE,
  threshold = 0,
  lim = min(max(c(r2dz.x, 0.4), na.rm = TRUE) + 0.1, 0.5),
  legend = TRUE,
  legend.title = NULL,
  legend.labels = NULL,
  cex.legend = 0.65,
  legend.bty = "n",
  mark.x = NULL,
  mark.label = NULL,
  col.mark = "blue",
  xlab = NULL,
  ylab = NULL,
  cex.lab = 0.7,
  cex.axis = 0.7,
  list.par = list(oma = c(1, 1, 1, 1)),
  ...
)
}
\arguments{
\item{...}{arguments passed to other methods. First argument should either be an \code{\link{lm}} model with the
outcome regression, a \code{\link{formula}} describing the model along
with the \code{\link{data.frame}} containing the variables of the model,
or a numeric vector with the coefficient estimate.}

\item{model}{An \code{lm} or \code{fixest} object with the outcome regression.}

\item{treatment}{A character vector with the name of the treatment variable
of the model.}

\item{benchmark_covariates}{The user has two options: (i)  character vector of the names of covariates that will be used to bound the plausible strength of the unobserved confounders. Each variable will be considered separately; (ii) a named list with character vector names of covariates that will be used, \emph{as a group}, to bound the plausible strength of the unobserved confounders. The names of the list will be used for the benchmark labels. Note: for factor variables with more than two levels, you need to provide the name of each level as encoded in the \code{fixest} model (the columns of \code{model.matrix}).}

\item{kd}{numeric vector. Parameterizes how many times stronger the confounder is related to the treatment in comparison to the observed benchmark covariate.
Default value is \code{1} (confounder is as strong as benchmark covariate).}

\item{r2yz.dx}{hypothetical partial R2 of unobserved confounder Z
with outcome Y, given covariates X and treatment D.}

\item{r2dz.x}{hypothetical partial R2 of unobserved confounder
Z with treatment D, given covariates X.}

\item{reduce}{should the bias adjustment reduce or increase the
absolute value of the estimated coefficient? Default is \code{TRUE}.}

\item{threshold}{estimate threshold.}

\item{lim}{sets limit for both axis. If `NULL`, limits are computed automatically.}

\item{legend}{should legend be plotted? Default is \code{TRUE}.}

\item{cex.legend}{size of the text for the legend.}

\item{legend.bty}{legend box. See \code{bty} argument of \link{par}.}

\item{formula}{an object of the class \code{\link{formula}}: a symbolic description of the model to be fitted.}

\item{method}{the default is \code{\link{lm}}. This argument can be changed to estimate the model using \code{\link[fixest]{feols}}. In this case the formula needs to be written so it can be estimated with \code{\link[fixest]{feols}} and the package needs to be installed.}

\item{vcov}{the variance/covariance used in the estimation when using \code{\link[fixest]{feols}}. See \code{\link[fixest]{vcov.fixest}} for more details. Defaults to "iid".}

\item{data}{data needed only when you pass a formula as first parameter. An object of the class \code{\link{data.frame}} containing the variables used in the analysis.}

\item{estimate}{Coefficient estimate.}

\item{se}{standard error of the coefficient estimate.}

\item{dof}{residual degrees of freedom of the regression.}

\item{legend.title}{the legend title. If \code{NULL}, then default legend is used.}

\item{legend.labels}{character vector of labels for the \code{r2yz.dx} curves.
If \code{NULL} (default), the labels are the \code{r2yz.dx} values as percentages,
rounded to one decimal place.}

\item{mark.x}{if not \code{NULL}, the x position at which to mark a point on the
threshold line. Ignored unless it lies within the plotting limits.}

\item{mark.label}{optional text label for the point drawn at \code{mark.x}.}

\item{col.mark}{colour of the point drawn at \code{mark.x}.}

\item{xlab}{label of x axis. If `NULL`, default label is used.}

\item{ylab}{label of y axis. If `NULL`, default label is used.}

\item{cex.lab}{The magnification to be used for x and y labels relative to the current setting of cex.}

\item{cex.axis}{The magnification to be used for axis annotation relative to the current setting of cex.}

\item{list.par}{arguments to be passed to \code{\link{par}}. It needs to be a named list.}
}
\value{
The function returns invisibly the data used for the extreme plot.
}
\description{
Extreme scenario plots of omitted variable bias for sensitivity analysis. The main inputs are an \code{\link{lm}} model, the treatment variable
and the covariates used for benchmarking the strength of unobserved confounding.

The horizontal axis shows the partial R2 of the unobserved confounder(s) with the treatment. The vertical axis shows the adjusted treatment effect estimate.
The partial R2 of the confounder with the outcome is represented by \emph{different curves} for each scenario, as given by the parameter \code{r2yz.dx}.
The red marks on horizontal axis are bounds on the partial R2 of the unobserved confounder \code{kd} times as strong as the covariates used for benchmarking.
The dotted red line represent the threshold for the effect estimate deemed to be problematic (for instance, zero).

See Cinelli and Hazlett (2020) for details.
}
\examples{

# runs regression model
model <- lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
                         pastvoted + hhsize_darfur + female + village,
                         data = darfur)
# extreme scenarios plot
ovb_extreme_plot(model, treatment = "directlyharmed",
                        benchmark_covariates = "female",
                        kd = 1:2,
                        lim = 0.05)

}
\references{
Cinelli, C. and Hazlett, C. (2020), "Making Sense of Sensitivity: Extending Omitted Variable Bias." Journal of the Royal Statistical Society, Series B (Statistical Methodology).
}
