% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovb_plot.R
\name{ovb_plot.sensemakr}
\alias{ovb_plot.sensemakr}
\title{Traditional omitted variable bias plot}
\usage{
\method{ovb_plot}{sensemakr}(x, lim = c(0, 0.4, 0.01), nlevels = 20,
  col.contour = "grey40", col.line = "red", ...)
}
\arguments{
\item{x}{A `sensemakr` object.}

\item{lim}{A numeric vector of length 3 containing the beginning, end, and
increment of a sequence, forms the axes of the omitted variable bias plot}

\item{nlevels}{The number of contour lines to plot}

\item{col.contour}{A character string describing the color of the contour
lines}

\item{col.line}{A character string describing the color of the critical value
/threshold line}

\item{...}{Additional graphics parameters}
}
\description{
This function produces a traditional omitted variable bias plot, which plots
the impact of an omitted variable on the relationship on interest contingent
on imbalance between treatment and control groups and a relationship between
the hypothetical variable and the outcome. We do not recommend using this
function, but provide it for compatibility with previous approaches.
}
