% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{print.sensemakr}
\alias{print.sensemakr}
\title{print.sensemakr is the print method for sensemakr class}
\usage{
\method{print}{sensemakr}(x, str = TRUE, ...)
}
\arguments{
\item{x}{object with 'sensemakr' class.}

\item{str}{a logical (default TRUE) indicating if the structure should be printed.}

\item{...}{extra args}
}
\description{
the print method for sensemakr class
}
\examples{
# loads data
data("darfur")

# fits model
model  = lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
               pastvoted + hhsize_darfur + female + village, data = darfur)

# runs benchmarking etc
sense = sensemakr(model, treatment = "directlyharmed")

print(sense)
print(sense,str=FALSE)

}
