% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{print.summary.sensemakr}
\alias{print.summary.sensemakr}
\title{print.summary.sensemakr is the print method for the summary of sensemakr}
\usage{
\method{print}{summary.sensemakr}(x, q = 1, scenarios = c(1, 0.25), ...)
}
\arguments{
\item{x}{object with 'sensemakr' class.}

\item{q}{FOO}

\item{scenarios}{a numeric vector, wher each element represents a scenario}

\item{...}{extra args}
}
\description{
the print method for summary of sensemakr class
}
