% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_benchmarks.R
\name{remove_benchmarks}
\alias{remove_benchmarks}
\alias{remove_benchmarks.default}
\alias{remove_benchmarks.sensemakr}
\title{Remove benchmarks from sensemakr object}
\usage{
remove_benchmarks(x)

\method{remove_benchmarks}{default}(x)

\method{remove_benchmarks}{sensemakr}(x)
}
\arguments{
\item{x}{A `sensemakr` object to remove benchmarks from.}
}
\value{
The `sensemakr` object `x` with benchmarks removed
}
\description{
This function removes any benchmark variables attached to a `sensemakr`
object.
}
\examples{
# Creating a sensemakr object using the built-in `darfur` data
data(darfur)
sense.out = sensemakr(formula = peacefactor ~ directlyharmed + female +
                        village + age,
                      data = darfur,
                      treatment = "directlyharmed",
                      benchmark = "female")

# Removing benchmarks
sense.out = remove_benchmarks(sense.out)

}
