% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{resid_maker}
\alias{resid_maker}
\title{Creates orthogonal residuals}
\usage{
resid_maker(n, C)
}
\arguments{
\item{n}{sample size.}
}
\value{
The function returns a numeric vector orthogonal to C.
}
\description{
This function is an auxiliary function for simulation purposes.
It creates a vector of n standard normal random variables, residualizes
this vector against a matrix of covariates C, then standardizes the vector again.
}
