% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_stats.R
\name{robustness_value}
\alias{robustness_value}
\alias{robustness_value.lm}
\alias{robustness_value.numeric}
\title{Robustness value of a regression coefficient}
\usage{
robustness_value(...)

\method{robustness_value}{lm}(model, covariates = NULL, q = 1,
  alpha = NULL, ...)

\method{robustness_value}{numeric}(t_statistic, dof, q = 1, alpha = NULL,
  ...)
}
\arguments{
\item{...}{arguments passed to other methods. First argument should either be an \code{lm} model with the
regression model or a numeric vector with the t-value of the coefficient estimate}

\item{model}{an \code{lm} object with the regression model.}

\item{covariates}{model covariates for which the robustness value will be computed. Default is to compute
the robustness value of all covariates.}

\item{q}{hypothesized proportion of bias relative to original estimate due to unobserved confounders,
used for computation of the robustness value. It has to be greater than zero. Default is \code{1}.}

\item{alpha}{significance level used for computation of the robustness value. Default is \code{0.05}.}

\item{t_statistic}{\code{numeric} vector with the t-value of the coefficient estimates}

\item{dof}{residual degrees of freedom of the regression}
}
\description{
Robustness value of a regression coefficient
}
