% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensemakr_function.R
\name{sensemakr}
\alias{sensemakr}
\title{Create a sensemakr object}
\usage{
sensemakr(formula, treatment, data = NULL, benchmark = NULL,
  verbose = FALSE)
}
\arguments{
\item{formula}{A formula describing the relationship Y ~ D + X}

\item{treatment}{A quoted character string naming the treatment variable in
`formula`.}

\item{data}{A data.frame object to fit models to.}

\item{benchmark}{A quoted character string. If provided, selects a
benchmark variable to compare potential unobserved confounders to.}

\item{verbose}{A logical value, if TRUE will print additional debugging
information while compiling the sensemakr object}
}
\value{
A `sensemakr` object.
}
\description{
Create a sensemakr object
}
