% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_stats.R
\name{sensitivity_stats}
\alias{sensitivity_stats}
\title{test}
\usage{
sensitivity_stats(model, treatment, q = 1, alpha = 0.05)
}
\arguments{
\item{model}{test}

\item{treatment}{test}

\item{q}{test}

\item{alpha}{test}
}
\description{
test
}
