% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{partial_r2}
\alias{partial_r2}
\alias{partial_r2.numeric}
\alias{partial_r2.lm}
\alias{partial_f2}
\alias{partial_f2.numeric}
\alias{partial_f2.lm}
\alias{robustness_value}
\alias{robustness_value.numeric}
\alias{robustness_value.lm}
\alias{rv}
\title{The RV (robusntess value), partial R^2, and partial f^2}
\usage{
partial_r2(...)

\method{partial_r2}{numeric}(t_statistic, dof)

\method{partial_r2}{lm}(model, covariate = NULL)

partial_f2(...)

\method{partial_f2}{numeric}(t_statistic, dof)

\method{partial_f2}{lm}(model, covariate = NULL)

robustness_value(..., q = 1, alpha = NULL)

\method{robustness_value}{numeric}(t_statistic, dof, q = 1, alpha = NULL)

\method{robustness_value}{lm}(model, covariate = NULL, q = 1,
  alpha = NULL)

rv(..., q = 1, alpha = NULL)
}
\arguments{
\item{...}{Additional parameters, not currently used.}

\item{t_statistic}{In lieu of supplying `model` and `covariate`, supply the
t-statistic of the treatment effect and `dof`}

\item{dof}{In lieu of supplying `model` and `covariate`, supply the residual
degrees of freedom of a model and `t_statistic`.}

\item{model}{An `lm` object which will be used to produce a robustness value}

\item{covariate}{A quoted character string describing the name of the
treatment covariate in the model object.}

\item{q}{A numeric fraction between 0 and 1 describing q% attenuation of the
observed treatment effect, defaults to 1 (complete attenuation)}

\item{alpha}{If specified, denotes a (1 - alpha) confidence interval used to
calculate the RV. If specified, the RV will reflect the percentage of
residual variation in the treatment/outcome that must be explained in order
for a (1 - alpha) confidence interval about the effect to cross the critical
threshold implied by `q`. If left empty or NULL, the point estimate must
cross the critical threshold.}
}
\value{
A q-robustness value, which is a number from 0 to 1.
}
\description{
Statistics for OVB.

Calculates the RV (robustness value)
}
