% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{worstcaseinterpret}
\alias{worstcaseinterpret}
\title{a function to interpret sensemakr quantities}
\usage{
worstcaseinterpret(sensemakr, scenarios = c(1, 0.25), q = 1)
}
\arguments{
\item{sensemakr}{object with 'sensemakr' class.}

\item{scenarios}{a numeric vector where each element represents a R2 scenario}

\item{q}{numeric value between 0 and 1 representing proportion of alteration to treatment estimates}
}
\description{
provides text interpretation in the console
}
\examples{
# loads data
data("darfur")

# fits model
model  = lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
               pastvoted + hhsize_darfur + female + village, data = darfur)

# runs benchmarking etc
sense = sensemakr(model, treatment = "directlyharmed")

worstcaseinterpret(sense)

}
\seealso{
\code{\link{summary.sensemakr}}
}
