% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colombia}
\alias{colombia}
\title{Data from Colombia}
\format{A data frame with 1123 rows and 16 columns.
\describe{
  \item{dept_code}{}
  \item{department}{}
  \item{town_code}{}
  \item{town}{}
  \item{total_eligible}{}
  \item{yes_vote}{}
  \item{santos10}{}
  \item{santos14}{}
  \item{gdppc}{}
  \item{pop13}{}
  \item{elev}{}
  \item{fat_all}{}
  \item{fat_2001to2005_gtd}{}
  \item{fat_2006to2010_gtd}{}
  \item{fat_2011to2015_gtd}{}
  \item{fat_2010to2013}{}
}}
\usage{
colombia
}
\description{
Data on Colombia...

The main "treatment"
variables are \code{santos2014}, which indicates the share of ...
and \code{fat_2011to2015_gtd}, which indicates...

The main outcome of interest is \code{yes_vote}, a measure of...

Key  covariates include
\code{gdppc} (GDP per capita),
\code{elev} (elevation),
 \code{pop13} (population in ),
}
\examples{
loads data
data(colombia)

# Violence Models ---------------------------------------------------------

### Naive model --- Model 1
model1 <- lm(yes_vote ~ fat_2001to2005_gtd, data = colombia)

### Model 2
model2 <- lm(yes_vote ~ fat_2001to2005_gtd + fat_2006to2010_gtd +
               fat_2011to2015_gtd + total_eligible + santos10 + gdppc ,
             data = colombia)

### Sensitivity analysis (model 2)
sense.model2 <- sensemakr(model2,
                          treatment = "fat_2011to2015_gtd",
                          benchmark = "santos10",
                          kd = 1)

### contour plot point estimate
plot(sense.model2)

### contour plot t-value
plot(sense.model2, sensitivity.of = "t-value")


# Political Affiliation Model ---------------------------------------------

### Model 3
model3  <- lm(yes_vote ~ santos14 + fat_2010to2013 + elev + gdppc + pop13,
              data = colombia)


sense.model3 <- sensemakr(model3, treatment = "santos14",
                          benchmark = c("gdppc","elev"),
                          kd = 3)

### contour plot point estimate
plot(sense.model3, lim = .9)

### contour plot t-value
plot(sense.model3, sensitivity.of = "t-value", lim = 0.9)

}
\references{
Hazlett, Chad, and Francesca Parente. "Credible or Confounded? Applying sensitivity analyses to improve research and its evaluation under imperfect identification." (2018)
}
\keyword{datasets}
