% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_plot.sensemakr.R
\name{contour_plot.sensemakr}
\alias{contour_plot.sensemakr}
\title{Contour plot for `sensemakr` objects}
\usage{
\method{contour_plot}{sensemakr}(x, ..., lim = c(0, 0.4, 0.001),
  nlevels = 20, plot_t = FALSE, col.contour = "grey40",
  col.line = "red", multipliers_y = c(1, 2, 3),
  multipliers_d = multipliers_y)
}
\arguments{
\item{x}{A `sensemakr`  object to produce a contour plot from.}

\item{...}{Additional plotting graphical parameters}

\item{lim}{A numeric vector of length 3 containing the beginning, end, and
increment of a sequence, forms the axes of the bound plot}

\item{nlevels}{How many contour lines to produce}

\item{plot_t}{Logical, default FALSE describing whether to plot the
t-statistic (TRUE) or effect estimate (FALSE).}

\item{col.contour}{A color parameter for contour plot lines}

\item{col.line}{A color parameter for the threshold / critical value line}

\item{multipliers_y}{A vector of multipliers, k, describing how many times
as strong the hypothetical confounders are than the benchmark confounder
with respect to the outcome. Only usable with `sensemakr` objects that
contain benchmark variables.}

\item{multipliers_d}{A vector of multipliers, k, describing how many times
as strong the hypothetical confounders are than the benchmark confounder
with respect to the treatment. Only usable with `sensemakr` objects that
contain benchmark variables.}
}
\description{
Produces a sensitivity analysis contour plot. The produced plot maps how
observed effects could attenuate in the presence of an unobserved confounder
that is correlated with treatment and outcome.
}
\details{
By default, this plot will compare the unadjusted effect size estimate to
attenuated versions of the estimate under the presence of unobserved
confounding, using the benchmark variables specified in the `sensemakr`
object. The default comparison plots the effect under the presence of an
unobserved confounder at 1x, 2x, and 3x the "strength" of the benchmark
confounder. Users can override the benchmark confounders by specifying a
`benchmark_covariate` argument (a vector of character strings naming the
benchmarks to plot), and the multipliers by specifying `multipliers_y` and,
optionally, `mulitpliers_d`.
}
\examples{
# Creating a sensemakr object using the built-in `darfur` data
data(darfur)
sense.out = sensemakr(formula = peacefactor ~ directlyharmed + female +
                        village + age,
                      data = darfur,
                      treatment = "directlyharmed",
                      benchmark = "female")

# Basic contour plot
contour_plot(sense.out)

# Contour plot of t-values and critical threshold
contour_plot(sense.out, plot_t = TRUE)

# Overriding multipliers
contour_plot(sense.out, multipliers_y = c(1, 1.5))

# Overriding colors
contour_plot(sense.out, col.contour = "black", col.line = "blue")

# Adding graphical parameters
contour_plot(sense.out, main = "Custom Title Plot")

# Plotting only some of the benchmarks
sense.out = add_benchmark(sense.out, "age")
contour_plot(sense.out, benchmark_covariate = "female")


}
