% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_plot.default.R
\name{extreme_plot.default}
\alias{extreme_plot.default}
\title{Extreme scenario plot}
\usage{
\method{extreme_plot}{default}(estimate, se, dof, r2d = NULL, lim = NULL,
  scenarios = c(1, 0.8, 0.5), cex.legend = 0.5, ...)
}
\arguments{
\item{estimate}{An estimated effect magnitude}

\item{se}{The standard error of the estimated effect magnitude}

\item{dof}{The residual degrees of freedom of the regression producing the
effect estimate}

\item{r2d}{Partial R^2 with respect to the treatment}

\item{lim}{A numeric vector of length 3 containing the beginning, end, and
increment of a sequence, forms the axes of the bound plot}

\item{scenarios}{A vector of proportions of the residual variance explained
by the unobserved confounder. Defaults to `c(1, 0.8, 0.5)`}

\item{cex.legend}{A scaling factor for the legend text.}

\item{...}{Additional graphical parameters}
}
\description{
Produces an extreme scenario plot: a plot where unobserved confounders
explain all the left-out residual variance of the outcome.
}
