% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_plot.sensemakr.R
\name{extreme_plot.sensemakr}
\alias{extreme_plot.sensemakr}
\title{Extreme scenario plot}
\usage{
\method{extreme_plot}{sensemakr}(x, lim = NULL, scenarios = c(1, 0.8, 0.5),
  cex.legend = 0.5, ...)
}
\arguments{
\item{x}{A `sensemakr` object containing a benchmark covariate}

\item{lim}{A numeric vector of length 3 containing the beginning, end, and
increment of a sequence, forms the axes of the bound plot}

\item{scenarios}{A vector of proportions of the residual variance explained
by the unobserved confounder. Defaults to `c(1, 0.8, 0.5)`}

\item{cex.legend}{A scaling factor for the legend text.}

\item{...}{Additional graphical parameters}
}
\description{
Produces an extreme scenario plot: a plot where unobserved confounders
explain all the left-out residual variance of the outcome. Uses a `sensemakr`
object.
}
