% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensemakr.R
\name{get_bias}
\alias{get_bias}
\alias{get_bias}
\alias{get_se}
\alias{get_bias}
\alias{get_t}
\title{Computes effects on estimate, standard error and t-value caused by unobserved confounder}
\usage{
get_bias(se, df, r2y, r2d)

get_se(se, df, r2y, r2d)

get_t(t, df, r2y, r2d, reduce = TRUE)
}
\arguments{
\item{se}{standard error of original  treatment effect estimate}

\item{df}{degrees of freedom of the original linear model}

\item{r2y}{hypothetical partial R2 of the confounder with the outcome}

\item{r2d}{hypothetical partial R2 of the confounder with the treatment}

\item{...}{extra arguments}
}
\description{
These functions compute the bias caused by an unobserved confounder with a specific pair
of partial R2 with the treatment and with the outcome.
with certain characteristics.
}
