% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{interpret}
\alias{interpret}
\title{a function to interpret sensemakr quantities}
\usage{
interpret(sensemakr, q = 1)
}
\arguments{
\item{sensemakr}{a 'sensemakr' object, result of \code{\link{sensemakr}}}

\item{q}{numeric value between 0 and 1 representing proportion of alteration to treatment estimates}
}
\description{
provides text interpretation in the console
}
\examples{
# loads data
data("darfur")

# fits model
model  = lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
               pastvoted + hhsize_darfur + female + village, data = darfur)

# runs benchmarking etc
sense = sensemakr(model, treatment = "directlyharmed")

interpret(sense, q = 0.5)
interpret(sense, q = 0.6)
}
\seealso{
\code{\link{summary.sensemakr}} \code{\link{sensemakr}}
}
