% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{interpret}
\alias{interpret}
\title{Low level function to interpret sensemakr quantities}
\usage{
interpret(sensemakr, q = 1)
}
\arguments{
\item{sensemakr}{object with 'sensemakr' class.}

\item{q}{FOO}
}
\description{
provides text interpretation in the console
}
