% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ovb_extreme_plot}
\alias{ovb_extreme_plot}
\alias{ovb_extreme_plot.numeric}
\alias{ovb_extreme_plot.lm}
\alias{ovb_extreme_plot.formula}
\alias{ovb_extreme_plot.sensemakr}
\title{Extreme scenario plots}
\usage{
ovb_extreme_plot(...)

\method{ovb_extreme_plot}{numeric}(estimate, se, dof, r2dz.x = NULL,
  r2yz.dx = c(1, 0.75, 0.5), reduce = TRUE, threshold = 0,
  lim = min(c(r2dz.x + 0.1, 0.5)), cex.legend = 0.5, ...)

\method{ovb_extreme_plot}{lm}(model, treatment, benchmark_covariates = NULL,
  kd = 1, r2yz.dx = c(1, 0.75, 0.5), r2dz.x = NULL, reduce = TRUE,
  threshold = 0, lim = min(c(r2dz.x + 0.1, 0.5)), cex.legend = 0.5, ...)

\method{ovb_extreme_plot}{formula}(formula, data, treatment,
  benchmark_covariates = NULL, kd = 1, r2yz.dx = c(1, 0.75, 0.5),
  r2dz.x = NULL, reduce = TRUE, threshold = 0, lim = min(c(r2dz.x + 0.1,
  0.5)), cex.legend = 0.5, ...)

\method{ovb_extreme_plot}{sensemakr}(x, lim = NULL, scenarios = c(1, 0.75,
  0.5), cex.legend = 0.5, ...)
}
\arguments{
\item{...}{Arguments which will pass through to the methods invoked.}

\item{estimate}{An estimated effect magnitude}

\item{se}{The standard error of the estimated effect magnitude}

\item{dof}{The residual degrees of freedom of the regression producing the
effect estimate}

\item{lim}{A numeric vector of length 3 containing the beginning, end, and
increment of a sequence, forms the axes of the bound plot}

\item{cex.legend}{A scaling factor for the legend text.}

\item{x}{A `sensemakr` object containing a benchmark covariate}

\item{scenarios}{A vector of proportions of the residual variance explained
by the unobserved confounder. Defaults to `c(1, 0.8, 0.5)`}

\item{r2d}{Partial R^2 with respect to the treatment of a known, observed
confounder. This is used to produce a rug inscription on the x-axis in order
to benchmark the unobserved confounder versus known confounders.}

\item{...}{Additional graphical parameters}

\item{lim}{A numeric vector of length 3 containing the beginning, end, and
increment of a sequence, forms the axes of the bound plot}

\item{scenarios}{A vector of proportions of the residual variance explained
by the unobserved confounder. Defaults to `c(1, 0.8, 0.5)`}

\item{cex.legend}{A scaling factor for the legend text.}

\item{...}{Additional graphical parameters}
}
\description{
`extreme_plot` is a generic function to produce extreme scenario plots. The
function invokes particular methods which depend on the class of the first
argument. For documentation on using  `extreme_plot` with `sensemakr`
objects, check \link{extreme_plot.sensemakr}. For documentation on using
`extreme_plot` with direct numeric arguments, check
\link{extreme_plot.default}

Produces an extreme scenario plot: a plot where unobserved confounders
explain all the left-out residual variance of the outcome. The x-axis
represents increasing partial R^2 of the unobserved covariate with the
treatment. Users can use this plot to make claims that unobserved
confounders that explain the entire residual variance are unlikely, so
persistence of effect magnitudes or signs even in the presence of them
suggests an overall robust effect.

Produces an extreme scenario plot: a plot where unobserved confounders
explain all the left-out residual variance of the outcome. The x-axis
represents increasing partial R^2 of the unobserved covariate with the
treatment. Users can use this plot to make claims that unobserved
confounders that explain the entire residual variance are unlikely, so
persistence of effect magnitudes or signs even in the presence of them
suggests an overall robust effect.
}
\details{
The easiest way to use `extreme_plot` is by creating a `sensemakr` object
and using the \link{extreme_plot.sensemakr} function on it. This mode of the
function is for users who want to directly pass estimates.

This mode of the function uses  `sensemakr` objects to produce the plot.
Users who wish to directly supply effect estimates manually should view
documentation on \link{extreme_plot.default}
}
\examples{
# Basic extreme scenario plot
extreme_plot(estimate = 0.09589905, se = 0.02318221, dof = 783,
             r2d = 0.008106797)

# Creating a sensemakr object using the built-in `darfur` data
data(darfur)
sense.out = sensemakr(formula = peacefactor ~ directlyharmed + female +
                        village + age,
                      data = darfur,
                      treatment = "directlyharmed",
                      benchmark = "female")

# Basic extreme scenario plot
extreme_plot(sense.out)

}
