% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_rv_r2_f2.R
\name{partial_r2}
\alias{partial_r2}
\alias{partial_r2.numeric}
\alias{partial_r2.lm}
\alias{group_partial_r2}
\alias{group_partial_r2.numeric}
\alias{group_partial_r2.lm}
\title{The RV (robustness value), partial R^2, and partial f^2}
\usage{
partial_r2(...)

\method{partial_r2}{numeric}(t_statistic, dof)

\method{partial_r2}{lm}(model, covariates = NULL)

group_partial_r2(...)

\method{group_partial_r2}{numeric}(f_statistic, p, dof)

\method{group_partial_r2}{lm}(model, covariates)
}
\description{
Statistics for OVB.
}
