% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.sensemade}
\alias{plot.sensemade}
\title{Sensitivity plots}
\usage{
\method{plot}{sensemade}(x, type = c("contour", "worst-case"), ...)
}
\arguments{
\item{x}{sensemade object}

\item{type}{type of the plot}

\item{...}{extra arguments}
}
\value{
a ggplot object with the plot.
}
\description{
Several sensitivity plots.
}

