% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.sensemakr}
\alias{plot.sensemakr}
\title{Plots for sensitivity analysis to unobserved confounders}
\usage{
\method{plot}{sensemakr}(x, type = "contour", ...)
}
\arguments{
\item{x}{A sensemakr object}

\item{type}{A character string reading "contour", "extreme", or "ovb".}

\item{...}{Further plotting arguments described in documentation for
\link{contour_plot.sensemakr}, \link{extreme_plot.sensemakr}, or
\link{ovb_plot.sensemakr}.}
}
\description{
This is a generic method which dispatches to various plotting functions for
`sensemakr` objects. By default, the plotting function chosen is
\link{contour_plot.sensemakr}. Using the `type` argument, users can also
select \link{extreme_plot.sensemakr} and \link{ovb_plot.sensemakr} plots.
}
\examples{
# Creating a sensemakr object using the built-in `darfur` data
data(darfur)
sense.out = sensemakr(formula = peacefactor ~ directlyharmed + female +
                        village + age,
                      data = darfur,
                      treatment = "directlyharmed",
                      benchmark = "female")

# Contour plot
plot(sense.out)
plot(sense.out, type="contour")

# Extreme Scenario plot
plot(sense.out, type="extreme")

# Traditional, unscaled OVB plot
plot(sense.out, type="ovb")

}
