% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.sensemakr}
\alias{plot.sensemakr}
\alias{plot.sensemakr}
\alias{contourplot}
\title{Sensitivity plots}
\usage{
\method{plot}{sensemakr}(x, showvars = "masked", type = c("contour",
  "worst-case"), ...)

contourplot(x, contour = c("estimate", "t-value", "lower-limit",
  "upper-limit"), nlevels = 15, pch = 20, cex = 1, lim = NULL,
  xlab = "Hypothetical partial R2 of unobserved confounder with the treatment",
  ylab = "Hypothetical partial R2 of unobserved confounder with the outcome",
  main = paste("Sensitivity of", contour,
  "to unobserved confounder\\nContours of adjusted estimates"), top = NULL)
}
\arguments{
\item{x}{sensemakr object}

\item{showvars}{chooses which subsets of benchmarks to display.
Valid options are: 'masked', 'all', or list('foo1','foo2')}

\item{type}{a character string representing the type of plot: "contour" or "worst-case".}

\item{...}{Arguments to be passed to methods, such as graphical parameters \code{\link{par}} .}

\item{contour}{a character string choosing what the contour lines represent: "estimate","t-value", "lower-limit", or "upper-limit"}

\item{nlevels}{an integer representing how many contour levels to display}

\item{pch}{see \code{\link{par}}}

\item{cex}{see \code{\link{par}}}

\item{lim}{a single numeric specifying the limits of a square plot window (one numeric applied to both x and y)}

\item{xlab}{see \code{\link{par}}}

\item{ylab}{see \code{\link{par}}}

\item{main}{see \code{\link{par}}}

\item{top}{an integer controlling the number of 'top' benchmarks to display}
}
\value{
a plot object
}
\description{
Several sensitivity plots.
}
\examples{
# loads data
data("darfur")

# fits model
model  = lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
               pastvoted + hhsize_darfur + female + village, data = darfur)

# runs benchmarking
sense = sensemakr(model, treatment = "directlyharmed")

# draws plots
plot1_data = plot(sense)
plot2_data = plot(sense, contour = "t-value")
plot3_data = plot(sense, contour = "lower-limit")
plot4_data = plot(sense, contour = "upper-limit")
plot5_data = plot(sense, type = "worst-case")


}
\seealso{
\code{\link{par}} for graphical parameters
}
