% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.sensemakr}
\alias{plot.sensemakr}
\title{Sensitivity plots}
\usage{
\method{plot}{sensemakr}(x, showvars = "masked", type = c("contour",
  "worst-case"), ...)
}
\arguments{
\item{x}{sensemakr object}

\item{showvars}{chooses which subsets of benchmarks to display
Valid options are: 'masked', 'all', or list('foo1','foo2')}

\item{type}{type of the plot}

\item{...}{extra arguments}
}
\value{
a ggplot object with the plot.
}
\description{
Several sensitivity plots.
}
