% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.sensemakr}
\alias{plot.sensemakr}
\alias{plot.sensemakr}
\alias{contourplot}
\alias{plot.sensemakr}
\alias{worstcaseplot}
\title{The plot method for a sensemakr object}
\usage{
\method{plot}{sensemakr}(x, type = c("contour", "worst-case"), ...)

contourplot(x, showvars = "masked", contour = c("estimate", "t-value",
  "lower-limit", "upper-limit"), nlevels = 15, lim = NULL, top = NULL,
  ptlab = TRUE, pch = 20, cex = 1,
  xlab = "Hypothetical partial R2 of unobserved confounder with the treatment",
  ylab = "Hypothetical partial R2 of unobserved confounder with the outcome",
  main = paste("Sensitivity of", contour,
  "to unobserved confounder\\nContours of adjusted estimates"))

worstcaseplot(x, showvars = "masked", lim = NULL, scenarios = c(1, 0.3),
  cex.legend = 0.6,
  xlab = "Hypothetical partial R2 of unobserved confounder(s) with treatment",
  ylab = "Adjusted estimate",
  main = paste("Sensitivity of estimate to unobservedconfounder(s)\\n",
  "\\"Worst-case\\" scenarios of partial R2 with outcome"))
}
\arguments{
\item{x}{a `sensemakr` object, result of \code{\link{sensemakr}}}

\item{type}{a character string representing the type of plot: "contour" or "worst-case".}

\item{...}{extra arguments that might be passed to underlying functions}

\item{showvars}{chooses which subsets of benchmarks to display.
Valid options are: 'masked', 'all', or list('foo1','foo2') see \code{\link{list}}}

\item{contour}{a character string choosing what the contour lines represent: "estimate","t-value", "lower-limit", or "upper-limit"}

\item{nlevels}{an integer representing how many contour levels to display}

\item{lim}{a single numeric specifying the limits of a square plot-window (one numeric applied to both x and y)}

\item{top}{an integer controlling the number of 'top' ranked benchmarks to display}

\item{ptlab}{a logical (default TRUE) controlling if points are labeled}

\item{pch}{see \code{\link{par}}}

\item{cex}{see \code{\link{par}}}

\item{xlab}{see \code{\link{par}}}

\item{ylab}{see \code{\link{par}}}

\item{main}{see \code{\link{par}}}

\item{scenarios}{a numeric vector where each element represents a worst-case scenario for R2}

\item{cex.legend}{a seperate `cex` argument used for drawing the legend. See \code{\link{par}}}
}
\description{
This generic plot method can create several sensitivity analysis plots.
}
\examples{
# loads data
data("darfur")

# fits model
model  = lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
               pastvoted + hhsize_darfur + female + village, data = darfur)

# runs benchmarking
sense = sensemakr(model, treatment = "directlyharmed")

# draws plots
plot1_data = plot(sense)
plot2_data = plot(sense, contour = "t-value")
plot(sense, contour = "lower-limit")
plot(sense, contour = "upper-limit")
plot(sense, type = "worst-case")

# also see vignette(sensemakr) for further point label customization

# calling contourplot() directly instead of using the generic plot()
contourplot(sense,showvars='all')

# calling worstcaseplot() directly instead of using the generic plot()
worstcaseplot(sense,showvars='all')
}
\seealso{
See \code{\link{par}} for graphical parameters
and \code{\link{plot}} for the generic plot method
}
