% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{print.sensemakr}
\alias{print.sensemakr}
\title{print.sensemakr is the print method for sensemakr class}
\usage{
\method{print}{sensemakr}(x, str = TRUE, ...)
}
\arguments{
\item{x}{object with 'sensemakr' class.}

\item{str}{a logical (default TRUE) indicating if the structure should be printed.}

\item{...}{extra args}
}
