% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.sensemakr}
\alias{print.sensemakr}
\title{Print a sensemakr object}
\usage{
\method{print}{sensemakr}(x, print_covariates = FALSE, sort_by = "default",
  digits = max(3L, getOption("digits") - 2L), ...)
}
\arguments{
\item{x}{A `sensemakr` object}

\item{print_covariates}{Logical, describing whether to print the covariates
used to produce the sensemakr object. Defaults to FALSE.}

\item{sort_by}{A character string specifying how to sort the covariates, if
printed. Valid options are "default" ("model", which sorts by the order the
variables were specified in the formula generating the sensemakr object),
"model", "alpha" (alphabetical sort), "R2D" (sorts by R^2 with respect to
treatment, descending), and "R2Y" (sorts by R^2 with respect to outcome,
descending)}

\item{digits}{How many digits to round output to. Default is the greater of
3 digits or getOption("digits") - 2.}

\item{...}{Further arguments passed to print, not currently used.}
}
\description{
This function prints a sensemakr object, including all partial R^2 values for
covariates used to generate the sensemakr object, by default sorting by
variables in the order they were specified by the supplying formula.
}
\details{
Additional printing parameters include `digits` (specifying how many digits
of precision to include in numeric tables), and `sort_by` (a character
string specifying how to sort covariates printed; valid values include
"default", "alpha", "R2D", and "R2Y").
}
