% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.sensemakr}
\alias{print.sensemakr}
\title{test}
\usage{
\method{print}{sensemakr}(x, digits = max(3L, getOption("digits") - 2L), ...)
}
\arguments{
\item{x}{test}

\item{digits}{test}

\item{...}{test}
}
\description{
test
}
