% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{print.sensemakr}
\alias{print.sensemakr}
\title{The print method for a sensemakr object}
\usage{
\method{print}{sensemakr}(x, str = TRUE, ...)
}
\arguments{
\item{x}{a `sensemakr` object, result of \code{\link{sensemakr}}}

\item{str}{a logical (default TRUE) indicating if the structure should be printed.}

\item{...}{extra arguments that might be passed to underlying functions}
}
\description{
provides text interpretation in the console
}
\examples{
# loads data
data("darfur")

# fits model
model  = lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
               pastvoted + hhsize_darfur + female + village, data = darfur)

# runs benchmarking etc
sense = sensemakr(model, treatment = "directlyharmed")

print(sense)
print(sense,str=FALSE)

}
\seealso{
\code{\link{print}}
}
