% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{print.summary.sensemakr}
\alias{print.summary.sensemakr}
\title{The print method for the summary of a sensemakr object}
\usage{
\method{print}{summary.sensemakr}(x, q = 1, scenarios = c(1, 0.25), ...)
}
\arguments{
\item{x}{a `sensemakr` object, result of \code{\link{sensemakr}}}

\item{q}{numeric value between 0 and 1 representing proportion of alteration to treatment estimates}

\item{scenarios}{a numeric vector where each element represents a R2 scenario}

\item{...}{extra arguments that might be passed to underlying functions}
}
\examples{
# loads data
data("darfur")

# fits model
model  = lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
               pastvoted + hhsize_darfur + female + village, data = darfur)

# runs benchmarking etc
sense = sensemakr(model, treatment = "directlyharmed")

print(summary(sense))

}
\seealso{
\code{\link{print}}
}
