% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_rv_r2_f2.R
\name{robustness_value}
\alias{robustness_value}
\alias{robustness_value.numeric}
\alias{robustness_value.lm}
\alias{rv}
\title{Calculates the RV (robustness value)}
\usage{
robustness_value(..., q = 1, alpha = NULL)

\method{robustness_value}{numeric}(t_statistic, dof, q = 1, alpha = NULL,
  ...)

\method{robustness_value}{lm}(model, covariates = NULL, q = 1,
  alpha = NULL, ...)

rv(..., q = 1, alpha = NULL)
}
\arguments{
\item{...}{Additional parameters, not currently used.}

\item{q}{A numeric fraction between 0 and 1 describing q% attenuation of the
observed treatment effect, defaults to 1 (complete attenuation)}

\item{alpha}{If specified, denotes a (1 - alpha) confidence interval used to
calculate the RV. If specified, the RV will reflect the percentage of
residual variation in the treatment/outcome that must be explained in order
for a (1 - alpha) confidence interval about the effect to cross the critical
threshold implied by `q`. If left empty or NULL, the point estimate must
cross the critical threshold.}

\item{t_statistic}{In lieu of supplying `model` and `covariates`, supply the
t-statistic of the treatment effect and `dof`}

\item{dof}{In lieu of supplying `model` and `covariates`, supply the residual
degrees of freedom of a model and `t_statistic`.}

\item{model}{An `lm` object which will be used to produce a robustness value}

\item{covariates}{A quoted character string describing the name of the
treatment covariates in the model object.}
}
\value{
A q-robustness value, which is a number from 0 to 1.
}
\description{
Calculates the RV (robustness value)
}
