% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness_value.R
\name{robustness_value}
\alias{robustness_value}
\title{Calculates an RV (robustness value) statistic}
\usage{
robustness_value(model = NULL, covariate = NULL, q = 1, alpha = NULL,
  ..., t_statistic = NULL, dof = NULL)
}
\arguments{
\item{model}{An `lm` object which will be used to produce a robustness value}

\item{covariate}{A quoted character string describing the name of the
treatment covariate in the model object.}

\item{q}{A numeric fraction between 0 and 1 describing q% attenuation of the
observed treatment effect, defaults to 1 (complete attenuation)}

\item{alpha}{If specified, denotes a (1 - alpha) confidence interval used to
calculate the RV. If specified, the RV will reflect the percentage of
residual variation in the treatment/outcome that must be explained in order
for a (1 - alpha) confidence interval about the effect to cross the critical
threshold implied by `q`. If left empty or NULL, the point estimate must
cross the critical threshold.}

\item{...}{Additional parameters, not currently used.}

\item{t_statistic}{In lieu of supplying `model` and `covariate`, supply the
t-statistic of the treatment effect and `dof`}

\item{dof}{In lieu of supplying `model` and `covariate`, supply the residual
degrees of freedom of a model and `t_statistic`.}
}
\value{
A q-robustness value, which is a number from 0 to 1.
}
\description{
This function calculates the robustness value statistic described in "Make
Sense of Sensitivity". The robustness value describes the percentage of
residual variation in the treatment and outcome an unobserved confounder
would need to attenuate an observed effect size past a critical value
}
