% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensemakr.R
\docType{package}
\name{sensemakr-package}
\alias{sensemakr-package}
\title{sensemakr: sensitivity analysis, implementing "Making Sense of
Sensitivity: Extending Omitted Variable Bias"}
\description{
TODO: More documentation goes here
}
\references{
Cinelli, C. and Hazlett, C. "Making Sense of Sensitivity: Extending Omitted Variable Bias." (2018).
}
