% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensemakr.R
\docType{package}
\name{sensemakr-package}
\alias{sensemakr-package}
\title{sensemakr: A package for sensitivity analysis, implementing "Making Sense of
Sensitivity: Extending Omitted Variable Bias"}
\description{
More documentation goes here
}
