% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensemakr.R
\docType{package}
\name{sensemakr-package}
\alias{sensemakr-package}
\title{Sensitivity analysis: extending omitted variable bias"}
\description{
The sensemakr package implements a suite of sensitivity analysis tools that make it easier to
understand the impact of omitted variables in linear regression models.
It implements the methods discussed in Cinelli and Hazlett (2018).
}
\details{
The main function of the package is  \code{\link{sensemakr}}, which computes the most common sensitivity analysis results.
After running \code{sensemakr} you may directly use the plot and print methods in the returned object.

You may also find usefull to use some of the sensitivity functions directly, such as the sensitivity plots
(\code{\link{ovb_contour_plot}}, \code{\link{ovb_extreme_plot}}) the bias-adjusted estimates (\code{\link{adjusted_estimate}}),
the robustness value (\code{\link{robustness_value}}), covariates partial r2s (\code{\link{partial_r2}}), the bounds
on the strength of the confounders (\code{\link{ovb_bounds}}), among other convenience functions.

More information can be found on the help documentation, vignettes and related papers.
}
\references{
Cinelli, C. and Hazlett, C. "Making Sense of Sensitivity: Extending Omitted Variable Bias." (2018).
}
