% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensemakr.R
\name{sensemakr}
\alias{sensemakr}
\alias{sensemakr}
\alias{sensemakr.lm}
\title{The main function for senstivity analysis of linear models}
\usage{
sensemakr(model, treatment, group_list, ...)

\method{sensemakr}{lm}(model, treatment, group_list = NULL, ...)
}
\arguments{
\item{model}{the outcome model. Currently, only supports a formal \code{\link{lm}} object}

\item{treatment}{a character string with the treatment variable's name}

\item{group_list}{a list of character vectors where elements within one vector are terms that should be grouped}

\item{...}{extra arguments that might be passed to underlying functions}
}
\value{
The function returns an object of class 'sensemakr' which is a structured list containing
quantities for sensitivity analysis, namely:

\item{info}{list with information of the original outcome model
   \itemize{
      \item model the \code{\link{lm}} object for the original outcome model
      \item outcome the name of the outcome
      \item treatment the name of the treatment
      }
  }
\item{treat.stats}{list with the main statistics of the original treatment estimate which are needed for computing the sensitivity analysis.
   \itemize{
      \item df the degrees of freedom
      \item estimate the estimate of the treatment effect
      \item se the standard error of the estimate
      \item treat the name of the treatment
      }
  }
\item{benchmarks}{list of data frames with sensitivity analysis benchmarks
    \itemize{
    \item benchmark_dropallvar contains R2 benchmark values for worst-case scenarios.
    \item benchmark_eachvar contains R2 benchmark values for all columns of the outcome model's design matrix.
    \item benchmark_group contains R2 benchmark values for groups.
    \item benchmark_masked contains non-redundent R2 benchmark values useful for summaries
    \item benchmark_natural contains natural-scaled benchmark values.
    }
  }
}
\description{
This is the main function of the package. Users should run this function first to start their sensitivity analysis.
}
\examples{

# cleans workspace
rm(list = ls())

# library
library(sensemakr)

# loads data
data("darfur")

# fits model
model  = lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
               pastvoted + hhsize_darfur + female + village, data = darfur)

# runs benchmarking etc
sense = sensemakr(model, treatment = "directlyharmed")

print(sense)

# draws plots
plot(sense)

# text output
summary(sense)
summary(sense, q = 0.6)


}
\references{
Cite paper(s)
}
\seealso{
The supported methods for a `sensemakr` object are:
\code{\link{plot.sensemakr}} \code{\link{print.sensemakr}} \code{\link{summary.sensemakr}}.
Also see \code{\link{lm}}.
}
