% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensemakr.R
\name{sensemakr}
\alias{sensemakr}
\alias{sensemakr}
\alias{sensemakr.lm}
\title{Senstivity abalysis of linear models}
\usage{
sensemakr(...)

\method{sensemakr}{lm}(model, D, X = NULL, ...)
}
\arguments{
\item{...}{extra arguments}

\item{model}{the model.}

\item{D}{character vector with the treatment variable.}

\item{X}{character vector with the covariates for benchmarking.}
}
\value{
The function returns an object of class 'sensemakr' which is a list with the main
results for sensitivity analysis, namely:
\item{Treat Stats}{
   A list with the main statistics of the treatment estiamate which are needed for computing the sensitivity.
   \itemize{
      \item name
      \item estimate
      \item se
      \item df
      }
  }
\item{Benchmarks}{Benchmark values for the covariates listed on paramter X
    \itemize{
    \item R2 a data.frame with becnhmark values R2.
    \item SD a data.frame with benchmark values
    \item natural a data.frame with...
    }
  }
}
\description{
Description.
}
\examples{
# worked out example

}
\references{
Cite paper(s)
}
