% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensemakr.R
\name{sensemakr}
\alias{sensemakr}
\alias{sensemakr.lm}
\alias{sensemakr.formula}
\title{Sensitivity analysis to unobserved confounders}
\usage{
sensemakr(...)

\method{sensemakr}{lm}(model, treatment, benchmark_covariates = NULL,
  kd = 1, ky = kd, q = 1, alpha = 0.05, r2dz.x = NULL,
  r2yz.dx = r2dz.x, bound_label = "", reduce = TRUE, ...)

\method{sensemakr}{formula}(formula, data, treatment,
  benchmark_covariates = NULL, kd = 1, ky = kd, q = 1, alpha = 0.05,
  r2dz.x = NULL, r2yz.dx = r2dz.x, bound_label = "", reduce = TRUE, ...)
}
\arguments{
\item{...}{arguments passed to other methods. First argument should either be an \code{\link{lm}} model with the
outcome regression, or a \code{\link{formula}} describing the model along
with the \code{\link{data.frame}} containing the variables of the model.}

\item{model}{an \code{lm} object with the outcome regression.}

\item{treatment}{a character vector with the name of the treatment variable of the model.}

\item{benchmark_covariates}{a character vector of the names of covariates that will be used to bound the plausible strength
of the unobserved confounders.}

\item{kd}{numeric vector. Parameterizes how many times stronger the confounder is related to the treatment in comparison to the observed benchmark covariates.
Default value is \code{1}.}

\item{ky}{numeric vector. Parameterizes how many times stronger the confounder is related to the outcome in comparison to the observed benchmark covariates.
Default value is the same as \code{kd}.}

\item{q}{percent change of the effect estimate that would be deemed problematic.  Default is \code{1},
which means a reduction of 100\% of the current effect estimate (bring estimate to zero). It has to be greater than zero.}

\item{alpha}{significance level used for computation of the robustness value. If \code{NULL} (the default), the robustness value refers only to the point estimate, no sampling uncertainty is taken into account.}

\item{r2dz.x}{hypothetical partial R2 of unobserved confounder Z with treatment D, given covariates X.}

\item{r2yz.dx}{hypothetical partial R2 of unobserved confounder Z with outcome Y, given covariates X and treatment D.}

\item{bound_label}{label to bounds provided manually in \code{r2dz.x} and \code{r2yz.dx}.}

\item{reduce}{should the bias adjustment reduce or increase the absolute value of the estimated coefficient? Default is \code{TRUE}.}

\item{formula}{an object of the class \code{\link{formula}}: a symbolic description of the model to be fitted.}

\item{data}{data needed only when you pass a formula as first parameter. An object of the class \code{\link{data.frame}} containing the variables used in the analysis.}
}
\description{
Sensitivity analysis to unobserved confounders
}
\examples{
# loads dataset
data("darfur")

# runs regression model
model <- lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
                         pastvoted + hhsize_darfur + female + village, data = darfur)

# runs sensemakr for sensitivity analysis
sensitivity <- sensemakr(model, treatment = "directlyharmed",
                               benchmark_covariates = "female",
                               kd = 1:3)

robustness_value(sensitivity)

# plot bias contour of point estimate
plot(sensitivity)

# plot bias contour of t-value
plot(sensitivity, sensitivity.of = "t-value")

# plot extreme scenario
plot(sensitivity, type = "extreme")

# v

}
\references{
Cinelli, C. and Hazlett, C. "Making Sense of Sensitivity: Extending Omitted Variable Bias." (2018).
}
