% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensemakr.R
\name{sensemakr}
\alias{sensemakr}
\alias{sensemakr}
\alias{sensemakr.lm}
\title{Senstivity analysis of linear models}
\usage{
sensemakr(model, treatment, benchmarks, ...)

\method{sensemakr}{lm}(model, treatment, group_list = NULL)
}
\arguments{
\item{model}{the model.}

\item{treatment}{character vector with the treatment variable.}

\item{group_list}{a list of character vectors where elements within one vector are terms that should be grouped.}
}
\value{
The function returns an object of class 'sensemakr' which is a list with the main
results for sensitivity analysis, namely:
\item{Treat Stats}{
   A list with the main statistics of the treatment estiamate which are needed for computing the sensitivity.
   \itemize{
      \item name
      \item estimate
      \item se
      \item df
      }
  }
\item{Benchmarks}{Benchmark values for the covariates listed on paramter X
    \itemize{
    \item R2 a data.frame with becnhmark values R2.
    \item SD a data.frame with benchmark values
    \item natural a data.frame with...
    }
  }
}
\description{
Description.
}
\examples{
rm(list = ls())

# library
library(sensemakr)

# loads data
data("darfur")

# fits model
model  <- lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
               pastvoted + hhsize_darfur + female + village, data = darfur)

# benchmark variables
X = c("herder_dar", "female", "age", "pastvoted", "farmer_dar")

# runs benchmarking etc
sense <- sensemakr(model, treatment = "directlyharmed", benchmarks = X)

# plots

## contour plot
plot1_data <- plot(sense)
plot2_data <- plot(sense, contour = "t-value")
plot3_data <- plot(sense, contour = "lower bound")
plot4_data <- plot(sense, contour = "upper bound")

## worst-case plot
plot5_data <- plot(sense, type = "worst-case")

# testing verbal outputs
interpret(sense)
interpret(sense, q = 0.6)

summary(sense)

}
\references{
Cite paper(s)
}
