% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_stats.R
\name{sensitivity_stats}
\alias{sensitivity_stats}
\alias{sensitivity_stats.lm}
\alias{sensitivity_stats.numeric}
\title{Sensitivity statistics for regression coefficients}
\usage{
sensitivity_stats(...)

\method{sensitivity_stats}{lm}(model, treatment, q = 1, alpha = 0.05, ...)

\method{sensitivity_stats}{numeric}(estimate, se, dof,
  treatment = "treatment", q = 1, alpha = 0.05, ...)
}
\arguments{
\item{...}{arguments passed to other methods. First argument should either be an \code{lm} model with the
outcome regression or a numeric vector with the coefficient estimate.}

\item{model}{an \code{lm} object with the outcome regression.}

\item{treatment}{a character vector with the name of the treatment variable of the model.}

\item{q}{percent change of the effect estimate that would be deemed problematic.  Default is \code{1},
which means a reduction of 100\% of the current effect estimate (bring estimate to zero). It has to be greater than zero.}

\item{alpha}{significance level used for computation of the robustness value. If \code{NULL} (the default), the robustness value refers only to the point estimate, no sampling uncertainty is taken into account.}

\item{estimate}{coefficient estimate.}

\item{se}{standard error of the coefficient estimate.}

\item{dof}{residual degrees of freedom of the regression.}
}
\value{
A \code{data.frame}.
}
\description{
Convenience function that computes the \code{\link{robustness_value}},
\code{\link{partial_r2}} and \code{\link{partial_f2}} of the coefficient of interest. It
returns a \code{data.frame} with all quantities of interest.
}
\examples{

## loads data
data("darfur")

## fits model
model <- lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
             pastvoted + hhsize_darfur + female + village, data = darfur)

## sensitivity stats for directly harmed
sensitivity_stats(model, treatment = "directlyharmed")

## you can  also pass the numeric values directly
sensitivity_stats(estimate = 0.09731582, se = 0.02325654, dof = 783)

}
