% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_stats.R
\name{sensitivity_stats}
\alias{sensitivity_stats}
\alias{sensitivity_stats.lm}
\alias{sensitivity_stats.numeric}
\title{Sensitivity statistics for regression coefficients}
\usage{
sensitivity_stats(...)

\method{sensitivity_stats}{lm}(model, treatment, q = 1, alpha = 0.05, ...)

\method{sensitivity_stats}{numeric}(
  estimate,
  se,
  dof,
  treatment = "treatment",
  q = 1,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{...}{Arguments passed to other methods. First argument should either be an
\code{lm} model with the outcome regression or a numeric vector with the
coefficient estimate.}

\item{model}{An \code{lm} object with the outcome regression.}

\item{treatment}{A character vector with the name of the treatment variable
of the model.}

\item{q}{percent change of the effect estimate that would be deemed problematic.  Default is \code{1},
which means a reduction of 100\% of the current effect estimate (bring estimate to zero). It has to be greater than zero.}

\item{alpha}{significance level used for computation of the robustness value. If \code{NULL} (the default), the robustness value refers only to the point estimate, no sampling uncertainty is taken into account (this is equivalent to set \code{alpha = 1}).}

\item{estimate}{Coefficient estimate.}

\item{se}{Standard error of the coefficient estimate.}

\item{dof}{Residual degrees of freedom of the regression.}
}
\value{
A \code{data.frame} containing the following quantities:
\describe{
\item{treatment}{a character with the name of the treatment variable}
\item{estimate}{a numeric vector with the estimated effect of the treatment}
\item{se}{a numeric vector with  the estimated standard error of the treatment effect}
\item{t_statistics}{a numeric vector with  the t-value of the treatment}
\item{r2yd.x}{a numeric vector with  the partial R2 of the treatment and the outcome, see details in \code{\link{partial_r2}}}
\item{rv_q}{a numeric vector with  the robustness value of the treatment, see details  in \code{\link{robustness_value}}}
\item{rv_qa}{a numeric vector with the robustness value of the treatment considering statistical significance, see details  in \code{\link{robustness_value}}}
\item{f2yd.x }{a numeric vector with the partial (Cohen's) f2 of the treatment with the outcome, see details in \code{\link{partial_f2}}}
\item{dof}{a numeric vector with the degrees of freedom of the model}
}
}
\description{
Convenience function that computes the \code{\link{robustness_value}},
\code{\link{partial_r2}} and \code{\link{partial_f2}} of the coefficient of interest.
}
\examples{

## loads data
data("darfur")

## fits model
model <- lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
             pastvoted + hhsize_darfur + female + village, data = darfur)

## sensitivity stats for directly harmed
sensitivity_stats(model, treatment = "directlyharmed")

## you can  also pass the numeric values directly
sensitivity_stats(estimate = 0.09731582, se = 0.02325654, dof = 783)

}
\references{
Cinelli, C. and Hazlett, C. (2020), "Making Sense of Sensitivity: Extending Omitted Variable Bias." Journal of the Royal Statistical Society, Series B (Statistical Methodology).
}
