% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{summary.sensemakr}
\alias{summary.sensemakr}
\title{summary.sensemakr is the summary method for sensemakr class}
\usage{
\method{summary}{sensemakr}(object, q = 1, scenarios = c(1, 0.25), ...)
}
\arguments{
\item{object}{with 'sensemakr' class.}

\item{q}{FOO}

\item{scenarios}{a numeric vector, wher each element represents a scenario}

\item{...}{extra args}
}
\description{
the summary method for sensemakr class
}
\examples{
# loads data
data("darfur")

# fits model
model  = lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
               pastvoted + hhsize_darfur + female + village, data = darfur)

# runs benchmarking etc
sense = sensemakr(model, treatment = "directlyharmed")

summary(sense)

}
