% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{worstcaseinterpret}
\alias{worstcaseinterpret}
\title{a Low level function to interpret sensemakr quantities}
\usage{
worstcaseinterpret(sensemakr, scenarios = c(1, 0.25), q = 1)
}
\arguments{
\item{sensemakr}{object with 'sensemakr' class.}

\item{scenarios}{a numeric vector, wher each element represents a scenario}

\item{q}{FOO}
}
